pyproject.toml
[project]
name = "docuxtract"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beanie>=1.26.0",
    "boto3>=1.35.29",
    "cryptography>=43.0.1",
    "fastapi[standard]>=0.115.0",
    "instructor>=1.4.3",
    "llama-index-embeddings-huggingface>=0.3.1",
    "llama-index-llms-mistralai>=0.2.5",
    "llama-index-llms-ollama>=0.3.2",
    "llama-index>=0.11.12",
    "motor>=3.6.0",
    "openai>=1.47.0",
    "opencv-python>=4.10.0.84",
    "pdf2image>=1.17.0",
    "pillow>=10.4.0",
    "pyjwt>=2.9.0",
    "pymongo>=4.9.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.1",
    "secure>=1.0.0",
    "spire-doc>=12.7.1",
]
